ov002: promote daObjFloatBoard_c into a single C++ TU
The abstract base of the planks that float on water -- Jolly Roger Bay's raft
and the two Wet-Dry World boards -- becomes one production source,
src/actors/daObjFloatBoard_c.cpp, absorbing five of the seven one-function
shards that used to carry it. The ROM build takes all five functions from a
single object.

WHAT THE CARTRIDGE PROVES ABOUT THE CLASS

  _ZTS  ov002 0x02108fc0  "17daObjFloatBoard_c"
  _ZTI  ov002 0x02108fb4  __si_class_type_info; +8 -> _ZTI10dBgActor_c
                          (ov002 0x021089ec), so the DIRECT base is
                          dBgActor_c and nothing else
  _ZTV  ov002 0x02108fdc  32 slots, 0x88 bytes of storage, the same count
                          as dBgActor_c -- this class adds no virtual
  size  0x348             the header's assert, corroborated by
                          daObjWcObj01_c's factory literal _ZN7fBase_cnwEj(840)

Every word was read straight out of extracted/overlays/overlay_0002.bin at
module base 0x020ad660, not out of src/.

THE CLASS IS ABSTRACT. Vtable slot 0 (InitResources) is a bare 0x00000000 with
no relocation and each of the three leaves supplies its own, which is why there
is no factory and no g_profile row. Five slots hold this class's own words: 3
CleanupResources, 6 Behavior, 9 Render, and the 16/17 destructor pair. Slot 31
still holds dBgActor_c::Kill, and every other slot still holds the inherited
arm9 word.

THE KEY FUNCTION IS CleanupResources, NOT Behavior. It is the first virtual
DECLARED out of line -- the destructor above it is inline in the class body and
the pure InitResources, declared last, has no body to be one. Both destructor
shards asserted Behavior; they were wrong about which function and right about
the consequence. This TU therefore emits the class's complete _ZTV/_ZTI/_ZTS
group, the whole inherited dBgActor_c -> dActor_c -> dBase_c -> fBase_c
vague-linkage RTTI chain, Vector3's inline destructor and both destructor
variants. None of that lies inside the licensed .text range, so dsd supplies
the cartridge's bytes and production isolation discards the copies; the
manifest licenses all FOURTEEN compiler-only outputs at their measured ROM
homes. tools/romdata_check.py --files src/actors/daObjFloatBoard_c.cpp reports
7 VERIFIED, 4 PARTIAL, 0 DIFFERS across the 11 data symbols.

WHY TWO SHARDS STAY. The cartridge places _ZN17daObjFloatBoard_cD0Ev at
0x020b5a18 BELOW _ZN17daObjFloatBoard_cD1Ev at 0x020b5a70, and mwccarm 2004/b56
will not emit that order from an inline-in-class destructor: measured here, the
vague-linkage pair always comes out D1-then-D0. A `delete p` scaffold, a
`p->~daObjFloatBoard_c()` scaffold and moving the destructor's declaration
below the three overrides were each tried and each produced D1-then-D0. The one
form that emits D0 first is an out-of-line definition -- it emits D2, D0, D1 --
and the cartridge rules that out: all three descendants INLINE this
destructor's vptr store (ov016 0x02112f3c/0x02112f9c/0x02112ff0 and ov029
0x021111e8/0x02111248/0x02111374 each load _ZTV17daObjFloatBoard_c at
0x02108fdc directly) and no instruction anywhere in the image calls 0x020b5a18
or 0x020b5a70 -- only vtable slots 17 and 16 name them. An out-of-line
destructor would make those three emit a call instead and break their already
matched TUs. So this entry licenses .text 0x020b5ab4..0x020b5e58,
src/_ZN17daObjFloatBoard_cD1Ev.cpp and src/_ZN17daObjFloatBoard_cD0Ev.cpp keep
their per-function entries, and this object's copies are licensed as
deadstrip-duplicate against those two ROM homes. Without that split the
pre-link audit refuses the whole TU with "licensed .text functions are not
emitted in ROM address order".

func_ov002_020b5ab4 and func_ov002_020b5b98 keep their address-derived names.
Neither occupies a vtable slot in any class in the image and nothing in the ROM
names them. 020b5ab4 has one inbound relocation, from this TU's Behavior;
020b5b98 has two, one from Behavior and one from func_ov002_020b5e58 in the
next TU, so it is not even file-local. Both keep their raw-offset bodies -- they
are the bytes that matched -- and Behavior now casts `this` at the two call
sites, because the merged TU cannot hold the shards' two contradictory
declarations of each name.

PROOF

  tools/tubuild.py verify ov002/daObjFloatBoard_c
    5/5 MATCH, objisolate clean, reloc-destinations clean -> TEXT-VERIFIED
    all 5 function(s) in the expected ROM-ascending section order
  tools/tubuild.py linkcheck ov002/daObjFloatBoard_c  (before promotion)
    SCRATCH-LINK-VERIFIED; TU .text range 0x020b5ab4..0x020b5e58 932 bytes
    IDENTICAL; ROM sha256 d1506e90efae5e2d2cf119926a4ac2a291bd5ca78349d09d5024e1a918c478e8,
    identical to the --baseline control; dsd check symbols --fail still fails
    on the same 9 pre-existing errors, 0 new
  tools/rombuild.py -j 6 --no-rom
    106/106 exact, 100.000000% of compared bytes, mismatching 0
    source-built functions 11,088, reproducing 11,088

  duplicate-sources, dead-references, tubuild-conflicts, check_src_tu,
  port_refcheck, layout_check, tu_order_check, cpp_tu_compat --require-ready
  and romdata_check all clean; converted-baseline re-banked with
  tiers_ratchet --update (2597 -> 2597, the two removals logged as the
  consolidation they are).

Also adds notes/data/class-facts/daObjFloatBoard_c.json, the scout stage's ROM
evidence for this class.
